Separate LogoAnimated
