+++ client/src/lib/components/Icons/Logo/LogoAnimated.web.tsx
@@ -1,0 +1,3 @@
+import React from 'react';
+
+export const LogoIconAnimated = () => null;
